splits the mappers file in different files
