+++ android/build.gradle
@@ -2,7 +2,7 @@
 
 buildscript {
     ext.android_tools = '4.0.0'
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.4.0'
     ext.android_junit5_version = '1.5.2.0'
     ext.junit5_runner = '0.2.2'
     ext.spek_version = '2.0.11'
@@ -77,9 +77,8 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.core:core-ktx:1.3.0'
-    api 'com.revenuecat.purchases:purchases:3.3.0'
+    implementation 'androidx.core:core-ktx:1.3.1'
+    api 'com.revenuecat.purchases:purchases:3.4.1'
     
     // assertion
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
