+++ android/src/main/java/com/revenuecat/purchases/common/mappers/EntitlementInfosMapper.kt
@@ -1,0 +1,9 @@
+package com.revenuecat.purchases.common.mappers
+
+import com.revenuecat.purchases.EntitlementInfos
+
+fun EntitlementInfos.map(): Map<String, Any> =
+    mapOf(
+        "all" to this.all.asIterable().associate { it.key to it.value.map() },
+        "active" to this.active.asIterable().associate { it.key to it.value.map() }
+    )
