+++ android/src/main/java/com/revenuecat/purchases/common/mappers/PurchasesPeriod.java
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.common;
+package com.revenuecat.purchases.common.mappers;
 
 import android.util.Log;
