+++ android/src/test/java/com/revenuecat/purchases/common/PeriodTests.kt
@@ -1,5 +1,6 @@
 package com.revenuecat.purchases.common
 
+import com.revenuecat.purchases.common.mappers.PurchasesPeriod
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
